+++ app/__init__.py
@@ -17,8 +17,10 @@ def create_app(config_filename=None):
     # Import and register Blueprints (for Controllers/Views)
     from .views.main import main as main_blueprint
     from .views.goods_and_categories import goods_and_categories as goods_and_categories_blueprint
+    from .views.staff_and_clients import staff_and_clients as staff_and_clients_blueprint
     app.register_blueprint(main_blueprint)
     app.register_blueprint(goods_and_categories_blueprint, url_prefix='/goods-and-categories')
+    app.register_blueprint(staff_and_clients_blueprint, url_prefix='/staff-and-clients')
 
     # Optionally, initialize Flask extensions like Flask-Login, Flask-Migrate, etc.
