+++ app/templates/components/_header.html
@@ -28,19 +28,11 @@
           </a>
         </li>
         <li>
-          <a href="{{ url_for('main.workers') }}" class="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 
+          <a href="{{ url_for('staff_and_clients.workers') }}" class="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 
           md:hover:bg-transparent md:border-0 md:hover:text-cyan-600 dark:text-white
           md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white 
           md:dark:hover:bg-transparent">
-           Our staff
-          </a>
-        </li>
-        <li>
-          <a href="#" class="block py-2 px-3 md:p-0 text-gray-900 rounded hover:bg-gray-100 
-          md:hover:bg-transparent md:border-0 md:hover:text-cyan-600 dark:text-white
-          md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white 
-          md:dark:hover:bg-transparent">
-           Our clients
+            Staff & Clients
           </a>
         </li>
         <li>
@@ -48,7 +40,7 @@
           md:hover:bg-transparent md:border-0 md:hover:text-cyan-600 dark:text-white
           md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white 
           md:dark:hover:bg-transparent">
-            Log in
+            Log in 
           </a>
         </li>
       </ul>
