--- app/templates/pages/workers.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet"  href="{{ url_for('static', filename='css/dist/styles.css') }}">
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/flowbite/2.3.0/flowbite.min.css" rel="stylesheet" />
-    <title>Our staff</title>
-</head>
-<body class="flex flex-col min-h-screen">
-    {% include 'components/_header.html' %}
-
-    
-    <div class="flex flex-col w-full pt-5 pb-8 items-center justify-center">
-        {% include 'components/_workers_topbar.html' %}
-
-        <section class="my-9 w-4/5 flex flex-wrap gap-6 justify-center">
-            {% include 'components/_worker_card.html' %}
-        </section>
-    </div>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/flowbite/2.3.0/flowbite.min.js"></script>
-</body>
-</html>
