+++ app/views/main.py
@@ -54,23 +54,10 @@ def goods_and_categories():
         user=user
         )
 
-@main.route('/workers')
-def workers():
-    positions = ["Manager", "Cashier", "Cleaner", "Consultant", "Security"]
-    workers = [
-        {
-            "name": f"Worker {i}",
-            "id": generate_unique_id(),
-            "position": random.choice(["Manager", "Cashier", "Cleaner", "Consultant", "Security"]),
-            "salary": f"${random.uniform(500, 900):.0f}",
-            "employment_date": f"2021-{random.randint(10, 12)}-{random.randint(10, 28)}",
-            "birth_date": f"{random.randint(1980, 2005)}-{random.randint(10, 12)}-{random.randint(10, 28)}",
-            "phone_number": generate_phone_number(),
-            "address": f"City, Street, {random.randint(1, 100)}",
-        } for i in range(1, 13)
-    ]
+@main.route('/staff-and-clients')
+def staff_and_clients():
     user = {'username': 'John Doe'} 
-    return render_template('pages/workers.html', positions=positions, workers=workers, user=user)
+    return render_template('pages/staff_and_clients.html', user=user)
 
 
 @main.route('/login')
